Relocate submit handlers
